Make the reqwest and futures feature orthogonal.
This way reqwest and futures may also be used without each other.
