+++ src/async_internal.rs
@@ -1,9 +1,6 @@
 use async_trait::async_trait;
 use failure::Fail;
 use futures::Future;
-use reqwest_futures_03::{Client, RedirectPolicy};
-
-use crate::reqwest::Error;
 use crate::{
     token_response, ClientCredentialsTokenRequest, CodeTokenRequest, ErrorResponse, HttpRequest,
     HttpResponse, PasswordTokenRequest, RefreshTokenRequest, RequestTokenError, TokenResponse,
@@ -187,44 +184,4 @@ where
             .map_err(RequestTokenError::Request)?;
         token_response(http_response)
     }
-}
-
-///
-/// Asynchronous HTTP client.
-///
-pub async fn async_http_client(request: HttpRequest) -> Result<HttpResponse, Error> {
-    let client = Client::builder()
-        // Following redirects opens the client up to SSRF vulnerabilities.
-        .redirect(RedirectPolicy::none())
-        .build()
-        .map_err(Error::Reqwest)?;
-
-
-    let mut request_builder = client
-        .request(request.method, request.url.as_str())
-        .body(request.body);
-    for (name, value) in &request.headers {
-        request_builder = request_builder.header(name, value);
-    }
-    let request = request_builder
-        .build()
-        .map_err(Error::Reqwest)?;
-
-    let response = client
-        .execute(request)
-        .await
-        .map_err(Error::Reqwest)?;
-
-    let status_code = response.status();
-    let headers = response.headers().clone();
-    let chunks = response
-        .bytes()
-        .await
-        .map_err(Error::Reqwest)?;
-
-     Ok(HttpResponse {
-        status_code,
-        headers,
-        body: chunks.to_vec(),
-    })
 }
