+++ src/reqwest.rs
@@ -1,25 +1,28 @@
+#[cfg(any(feature = "reqwest", feature = "reqwest-09"))]
 use std::io::Read;
 
 use failure::Fail;
-#[cfg(feature = "futures-01")]
+#[cfg(all(feature = "futures-01", feature = "reqwest-09"))]
 use futures_0_1::{Future, IntoFuture, Stream};
 
-#[cfg(feature = "futures-01")]
-use reqwest::r#async::Client as AsyncClient;
-#[cfg(feature = "futures-01")]
-use reqwest as blocking;
-#[cfg(feature = "futures-01")]
-use reqwest;
-#[cfg(feature = "futures-01")]
+#[cfg(all(feature = "futures-01", feature = "reqwest-09"))]
+use reqwest_09::r#async::Client as AsyncClient;
+#[cfg(feature = "reqwest-09")]
+use reqwest_09 as blocking;
+#[cfg(feature = "reqwest-09")]
+use reqwest_09::{RedirectPolicy};
+#[cfg(feature = "reqwest-09")]
+use reqwest_09 as reqwest;
+
+
+#[cfg(feature = "reqwest")]
+use reqwest::blocking;
+#[cfg(feature = "reqwest")]
 use reqwest::RedirectPolicy;
+#[cfg(all(feature = "futures-03", feature = "reqwest"))]
+use reqwest;
 
-#[cfg(feature = "futures-03")]
-use reqwest_futures_03::blocking;
-#[cfg(feature = "futures-03")]
-use reqwest_futures_03::RedirectPolicy;
-#[cfg(feature = "futures-03")]
-use reqwest_futures_03 as reqwest;
-
+#[cfg(any(feature = "reqwest", feature = "reqwest-09"))]
 use super::{HttpRequest, HttpResponse};
 
 ///
@@ -29,6 +32,7 @@ use super::{HttpRequest, HttpResponse};
 pub enum Error {
     /// Error returned by reqwest crate.
     #[fail(display = "request failed")]
+    #[cfg(any(feature = "reqwest", feature = "reqwest-09"))]
     Reqwest(#[cause] reqwest::Error),
     /// Non-reqwest HTTP error.
     #[fail(display = "HTTP error")]
@@ -44,6 +48,7 @@ pub enum Error {
 ///
 /// Synchronous HTTP client.
 ///
+#[cfg(any(feature = "reqwest", feature = "reqwest-09"))]
 pub fn http_client(request: HttpRequest) -> Result<HttpResponse, Error> {
     let client = blocking::Client::builder()
         // Following redirects opens the client up to SSRF vulnerabilities.
@@ -72,7 +77,7 @@ pub fn http_client(request: HttpRequest) -> Result<HttpResponse, Error> {
 ///
 /// Asynchronous HTTP client.
 ///
-#[cfg(feature = "futures-01")]
+#[cfg(all(feature = "futures-01", feature = "reqwest-09"))]
 pub fn async_http_client(request: HttpRequest) -> impl Future<Item = HttpResponse, Error = Error> {
     AsyncClient::builder()
     // Following redirects opens the client up to SSRF vulnerabilities.
@@ -109,5 +114,46 @@ pub fn async_http_client(request: HttpRequest) -> impl Future<Item = HttpRespons
                     })
                     .map_err(Error::Reqwest)
             })
+    })
+}
+
+///
+/// Asynchronous HTTP client.
+///
+#[cfg(all(feature = "reqwest", feature = "futures-03"))]
+pub async fn async_http_client(request: HttpRequest) -> Result<HttpResponse, Error> {
+    let client = reqwest::Client::builder()
+        // Following redirects opens the client up to SSRF vulnerabilities.
+        .redirect(RedirectPolicy::none())
+        .build()
+        .map_err(Error::Reqwest)?;
+
+
+    let mut request_builder = client
+        .request(request.method, request.url.as_str())
+        .body(request.body);
+    for (name, value) in &request.headers {
+        request_builder = request_builder.header(name, value);
+    }
+    let request = request_builder
+        .build()
+        .map_err(Error::Reqwest)?;
+
+    let response = client
+        .execute(request)
+        .await
+        .map_err(Error::Reqwest)?;
+
+    let status_code = response.status();
+    let headers = response.headers().clone();
+    let chunks = response
+        .bytes()
+        .await
+        .map_err(Error::Reqwest)?;
+
+     Ok(HttpResponse {
+        status_code,
+        headers,
+        body: chunks.to_vec(),
     })
 }
